chore: move branding type to types folder
