+++ src/previews/Preview6/Preview6.jsx
@@ -6,7 +6,7 @@ import colors from 'config/colors';
 import previewImg from 'assets/images/ip-preview-premium.png';
 
 // Types
-import { type Branding } from '../types';
+import { type Branding } from '../types/branding';
 
 // Components
 import TextEditor from '../components/TextEditor';
